# Conflicts:
#	.Makefile.lint
#	src/kernel_services/ast_printing/description.ml
#	src/kernel_services/ast_printing/logic_print.ml
#	src/kernel_services/ast_printing/printer_tag.ml
#	src/kernel_services/ast_printing/printer_tag.mli
#	src/libraries/utils/rich_text.ml
#	src/plugins/gui/design.ml
#	src/plugins/gui/history.ml
#	src/plugins/gui/pretty_source.ml
#	src/plugins/gui/pretty_source.mli
#	src/plugins/gui/property_navigator.ml
#	src/plugins/occurrence/register_gui.ml
#	src/plugins/scope/dpds_gui.ml
#	src/plugins/security_slicing/register_gui.ml
#	src/plugins/slicing/register_gui.ml
#	src/plugins/value/gui_files/register_gui.ml
#	src/plugins/wp/GuiSource.ml

+++ tests/value/volatile.c
@@ -80,7 +80,7 @@ int * volatile main2() {
   int * volatile p1, * volatile p2, * volatile p3;
   p1 = G ? 0 : &X;
   p2 = &X;
-  k = G ? 0 : &X;
+  k = G ? 0 : (int) &X;
   p3 = k;
   return k;
 }
